LayoutTests/media/replaceChild-should-not-pause-video-expected.txt

Test that a video element does not get paused when moved using replaceChild().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


EVENT(canplay)
video.play()
EVENT(playing)
PASS video.paused is false
parentA.replaceChild(video, childToReplace)
PASS video.paused is false
PASS video.paused is false
PASS successfullyParsed is true

TEST COMPLETE
